[Decomp][World] Cohesion split: extract config loading to WorldConfig.cpp
